services/requirements_login.txt
os
python-dotenv
flask
flask-dance
flask-sqlalchemy
flask-login
flask-cors
gunicorn
blinker
sqlalchemy-utils